Handle multisequence epitopes
A limitation of previous versions was the assumption that all epitopes came from the same sequence.
This made it overly complicated to work with epitopes from multiple sequences,
arguable a key feature.

Functions have been rewritten to take a table (instead of being applied to a table) and lookup the parent sequence separately for each epitope. This is done by the `ParentSeqSerialiser` class. It can apply the same parent sequence to all epitopes, lookup the parent sequence in a dictionary using the value of `parent_col` as a key, or take a sequence value from a column.

These function which take a table mostly work by applying a helper function to the table. The helper functions start with _ and are named to apply to a singular while the main function refers to the plural.

Function names and arguments have be made clearer and more cohesive as described in coordinate_system.