Simplify colouring and labelling plot_line() plots
For each row in the epiotpe table, the kwargs argument is updated based on the value of `row.color_col`.

For labelling, a dictionary of artists os maintained with a key for each value in `label_col`. The value of these keys is updated everytime a line is plotted, such that the last line plotted with a specific `label_col` value is the value in the dictionary. This means there is only one
key-value pair for each `label_col` value; therefore, each `label_col` value is only added once to the legend.
If there is no `label_col` value no legend is added but the dictionary of artists is still maintained, but with only a single key "".